+++ CHANGELOG.md
@@ -5,6 +5,9 @@ FEATURES:
 * add `host_group_ids` attribute in `yandex_mdb_kafka_cluster` resource and data source
 * add `host_group_ids` attribute in `yandex_dataproc_cluster` resource and data source
 
+ENHANCEMENTS:
+* add `dns_record`, `ipv6_dns_record` and `nat_dns_record` to `network_interface` entity in `yandex_compute_instance`
+
 ## 0.56.0 (April 15, 2021)
 ENHANCEMENTS:
 * dataproc: supported `security_group_ids`